Simplify NNAPI SharedHandle -- runtime
Prior to this CL, the NNAPI represented SharedHandle as a collection of
file descriptors and a vector of ints to act as metadata. This struct
mirrored Android's native_handle_t. However, the NNAPI only uses
SharedHandle to hold a single file descriptor to hold a single cache
file.

This CL simplifies the SharedHandle by changing it to be a single
ref-counted file descriptor.

Bug: N/A
Test: mma
Test: NeuralNetworksTest_static
Change-Id: If5d060182eee766ef067874844db0b0a13e686e9